move more separate unit tests into test_misc_features, ensure resources unit tests dont have race conditions by using temp directory
